Drop the concrete RestateEnvironment implementations in favor of a simple property holder
We now have a simple object to hold the necessary components of a "Restate environment pointer" – and its creation is hidden behind a factory method, so we can easily introduce new concrete implementations in the future, if it makes sense for specific deployment scenarios. In the meantime, this greatly simplifies the surface area at the cost of the user having to manually construct the admin URL for a given environment.